moved useAnimationFrame hook to hooks.tsx
